uplift state into container and make component dummy
